requirements/base.txt
html2text
requests
schedule
sh
pyyaml
selenium
xvfbwrapper
